Health_Index_Medpiper-main/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask_migrate
pandas
requests
OTPLessAuthSDK
flask-cors